scrapers/scrapy_scraper/requirements.txt
mysql-connector-python==8.2.0
numpy==1.24.4
beautifulsoup4==4.12.2
PyYAML==6.0.1
requests==2.31.0
requests-file==1.5.1
python-dotenv==1.0.0
Scrapy==2.11.0
scrapy-playwright==0.0.33
scrapyrt==0.15.0
service-identity==23.1.0
Twisted==22.10.0
